Network.Google.Resource.Calendar.ACL.Insert

Description

Creates an access control rule.

See: Calendar API Reference for calendar.acl.insert.

REST Resource

type ACLInsertResource = "calendar" :> ("v3" :> ("calendars" :> (Capture "calendarId" Text :> ("acl" :> (QueryParam "sendNotifications" Bool :> (QueryParam "alt" AltJSON :> (ReqBody '[JSON] ACLRule :> Post '[JSON] ACLRule)))))))

A resource alias for calendar.acl.insert method which the ACLInsert request conforms to.

Creating a Request

aclInsert

Arguments

:: Text
aiCalendarId

-> ACLRule
aiPayload

-> ACLInsert

Creates a value of ACLInsert with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aiCalendarId

aiPayload

aiSendNotifications

data ACLInsert

Creates an access control rule.

See: aclInsert smart constructor.

Request Lenses

aiCalendarId :: Lens' ACLInsert Text

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

aiPayload :: Lens' ACLInsert ACLRule

Multipart request metadata.

aiSendNotifications :: Lens' ACLInsert (Maybe Bool)

Whether to send notifications about the calendar sharing change. Optional. The default is True.