Network.Google.Resource.Calendar.ACL.Watch

Description

Watch for changes to ACL resources.

See: Calendar API Reference for calendar.acl.watch.

REST Resource

type ACLWatchResource = "calendar" :> ("v3" :> ("calendars" :> (Capture "calendarId" Text :> ("acl" :> ("watch" :> (QueryParam "syncToken" Text :> (QueryParam "showDeleted" Bool :> (QueryParam "pageToken" Text :> (QueryParam "maxResults" (Textual Int32) :> (QueryParam "alt" AltJSON :> (ReqBody '[JSON] Channel :> Post '[JSON] Channel)))))))))))

A resource alias for calendar.acl.watch method which the ACLWatch request conforms to.

Creating a Request

aclWatch

Arguments

:: Text
awCalendarId

-> Channel
awPayload

-> ACLWatch

Creates a value of ACLWatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

awSyncToken

awCalendarId

awShowDeleted

awPayload

awPageToken

awMaxResults

data ACLWatch

Watch for changes to ACL resources.

See: aclWatch smart constructor.

Request Lenses

awSyncToken :: Lens' ACLWatch (Maybe Text)

Token obtained from the nextSyncToken field returned on the last page of results from the previous list request. It makes the result of this list request contain only entries that have changed since then. All entries deleted since the previous list request will always be in the result set and it is not allowed to set showDeleted to False. If the syncToken expires, the server will respond with a 410 GONE response code and the client should clear its storage and perform a full synchronization without any syncToken. Learn more about incremental synchronization. Optional. The default is to return all entries.

awCalendarId :: Lens' ACLWatch Text

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

awShowDeleted :: Lens' ACLWatch (Maybe Bool)

Whether to include deleted ACLs in the result. Deleted ACLs are represented by role equal to "none". Deleted ACLs will always be included if syncToken is provided. Optional. The default is False.

awPayload :: Lens' ACLWatch Channel

Multipart request metadata.

awPageToken :: Lens' ACLWatch (Maybe Text)

Token specifying which result page to return. Optional.

awMaxResults :: Lens' ACLWatch (Maybe Int32)

Maximum number of entries returned on one result page. By default the value is 100 entries. The page size can never be larger than 250 entries. Optional.