Network.Google.Resource.Calendar.Calendars.Get

Description

Returns metadata for a calendar.

See: Calendar API Reference for calendar.calendars.get.

REST Resource

type CalendarsGetResource = "calendar" :> ("v3" :> ("calendars" :> (Capture "calendarId" Text :> (QueryParam "alt" AltJSON :> Get '[JSON] Calendar))))

A resource alias for calendar.calendars.get method which the CalendarsGet request conforms to.

Creating a Request

calendarsGet

Arguments

:: Text
cgCalendarId

-> CalendarsGet

Creates a value of CalendarsGet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cgCalendarId

data CalendarsGet

Returns metadata for a calendar.

See: calendarsGet smart constructor.

Request Lenses

cgCalendarId :: Lens' CalendarsGet Text

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.