Network.Google.Resource.Calendar.Calendars.Insert

Description

Creates a secondary calendar.

See: Calendar API Reference for calendar.calendars.insert.

REST Resource

type CalendarsInsertResource = "calendar" :> ("v3" :> ("calendars" :> (QueryParam "alt" AltJSON :> (ReqBody '[JSON] Calendar :> Post '[JSON] Calendar))))

A resource alias for calendar.calendars.insert method which the CalendarsInsert request conforms to.

Creating a Request

calendarsInsert

Arguments

:: Calendar
ciPayload

-> CalendarsInsert

Creates a value of CalendarsInsert with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciPayload

data CalendarsInsert

Creates a secondary calendar.

See: calendarsInsert smart constructor.

Request Lenses

ciPayload :: Lens' CalendarsInsert Calendar

Multipart request metadata.