Network.Google.Resource.Calendar.Events.Delete

Description

Deletes an event.

See: Calendar API Reference for calendar.events.delete.

REST Resource

type EventsDeleteResource = "calendar" :> ("v3" :> ("calendars" :> (Capture "calendarId" Text :> ("events" :> (Capture "eventId" Text :> (QueryParam "sendNotifications" Bool :> (QueryParam "sendUpdates" EventsDeleteSendUpdates :> (QueryParam "alt" AltJSON :> Delete '[JSON] ()))))))))

A resource alias for calendar.events.delete method which the EventsDelete request conforms to.

Creating a Request

eventsDelete

Arguments

:: Text
edCalendarId

-> Text
edEventId

-> EventsDelete

Creates a value of EventsDelete with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

edCalendarId

edSendNotifications

edSendUpdates

edEventId

data EventsDelete

Deletes an event.

See: eventsDelete smart constructor.

Request Lenses

edCalendarId :: Lens' EventsDelete Text

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

edSendNotifications :: Lens' EventsDelete (Maybe Bool)

Deprecated. Please use sendUpdates instead. Whether to send notifications about the deletion of the event. Note that some emails might still be sent even if you set the value to false. The default is false.

edSendUpdates :: Lens' EventsDelete (Maybe EventsDeleteSendUpdates)

Guests who should receive notifications about the deletion of the event.

edEventId :: Lens' EventsDelete Text

Event identifier.