Network.Google.Resource.Calendar.Events.Instances

Description

Returns instances of the specified recurring event.

See: Calendar API Reference for calendar.events.instances.

REST Resource

type EventsInstancesResource = "calendar" :> ("v3" :> ("calendars" :> (Capture "calendarId" Text :> ("events" :> (Capture "eventId" Text :> ("instances" :> (QueryParam "timeMin" DateTime' :> (QueryParam "showDeleted" Bool :> (QueryParam "originalStart" Text :> (QueryParam "maxAttendees" (Textual Int32) :> (QueryParam "pageToken" Text :> (QueryParam "timeZone" Text :> (QueryParam "maxResults" (Textual Int32) :> (QueryParam "alwaysIncludeEmail" Bool :> (QueryParam "timeMax" DateTime' :> (QueryParam "alt" AltJSON :> Get '[JSON] Events))))))))))))))))

A resource alias for calendar.events.instances method which the EventsInstances request conforms to.

Creating a Request

eventsInstances

Arguments

:: Text
eCalendarId

-> Text
eEventId

-> EventsInstances

Creates a value of EventsInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eCalendarId

eTimeMin

eShowDeleted

eOriginalStart

eMaxAttendees

ePageToken

eTimeZone

eMaxResults

eAlwaysIncludeEmail

eTimeMax

eEventId

data EventsInstances

Returns instances of the specified recurring event.

See: eventsInstances smart constructor.

Request Lenses

eCalendarId :: Lens' EventsInstances Text

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

eTimeMin :: Lens' EventsInstances (Maybe UTCTime)

Lower bound (inclusive) for an event's end time to filter by. Optional. The default is not to filter by end time. Must be an RFC3339 timestamp with mandatory time zone offset.

eShowDeleted :: Lens' EventsInstances (Maybe Bool)

Whether to include deleted events (with status equals "cancelled") in the result. Cancelled instances of recurring events will still be included if singleEvents is False. Optional. The default is False.

eOriginalStart :: Lens' EventsInstances (Maybe Text)

The original start time of the instance in the result. Optional.

eMaxAttendees :: Lens' EventsInstances (Maybe Int32)

The maximum number of attendees to include in the response. If there are more than the specified number of attendees, only the participant is returned. Optional.

ePageToken :: Lens' EventsInstances (Maybe Text)

Token specifying which result page to return. Optional.

eTimeZone :: Lens' EventsInstances (Maybe Text)

Time zone used in the response. Optional. The default is the time zone of the calendar.

eMaxResults :: Lens' EventsInstances (Maybe Int32)

Maximum number of events returned on one result page. By default the value is 250 events. The page size can never be larger than 2500 events. Optional.

eAlwaysIncludeEmail :: Lens' EventsInstances (Maybe Bool)

Whether to always include a value in the email field for the organizer, creator and attendees, even if no real email is available (i.e. a generated, non-working value will be provided). The use of this option is discouraged and should only be used by clients which cannot handle the absence of an email address value in the mentioned places. Optional. The default is False.

eTimeMax :: Lens' EventsInstances (Maybe UTCTime)

Upper bound (exclusive) for an event's start time to filter by. Optional. The default is not to filter by start time. Must be an RFC3339 timestamp with mandatory time zone offset.

eEventId :: Lens' EventsInstances Text

Recurring event identifier.