Network.Google.Resource.Calendar.Events.List

Description

Returns events on the specified calendar.

See: Calendar API Reference for calendar.events.list.

REST Resource

type EventsListResource = "calendar" :> ("v3" :> ("calendars" :> (Capture "calendarId" Text :> ("events" :> (QueryParam "syncToken" Text :> (QueryParam "timeMin" DateTime' :> (QueryParam "orderBy" EventsListOrderBy :> (QueryParam "singleEvents" Bool :> (QueryParams "privateExtendedProperty" Text :> (QueryParam "showDeleted" Bool :> (QueryParam "q" Text :> (QueryParams "sharedExtendedProperty" Text :> (QueryParam "maxAttendees" (Textual Int32) :> (QueryParam "iCalUID" Text :> (QueryParam "updatedMin" DateTime' :> (QueryParam "pageToken" Text :> (QueryParam "timeZone" Text :> (QueryParam "showHiddenInvitations" Bool :> (QueryParam "maxResults" (Textual Int32) :> (QueryParam "alwaysIncludeEmail" Bool :> (QueryParam "timeMax" DateTime' :> (QueryParam "alt" AltJSON :> Get '[JSON] Events))))))))))))))))))))))

A resource alias for calendar.events.list method which the EventsList request conforms to.

Creating a Request

eventsList

Arguments

:: Text
elCalendarId

-> EventsList

Creates a value of EventsList with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

elSyncToken

elCalendarId

elTimeMin

elOrderBy

elSingleEvents

elPrivateExtendedProperty

elShowDeleted

elQ

elSharedExtendedProperty

elMaxAttendees

elICalUId

elUpdatedMin

elPageToken

elTimeZone

elShowHiddenInvitations

elMaxResults

elAlwaysIncludeEmail

elTimeMax

data EventsList

Returns events on the specified calendar.

See: eventsList smart constructor.

Request Lenses

elSyncToken :: Lens' EventsList (Maybe Text)

Token obtained from the nextSyncToken field returned on the last page of results from the previous list request. It makes the result of this list request contain only entries that have changed since then. All events deleted since the previous list request will always be in the result set and it is not allowed to set showDeleted to False. There are several query parameters that cannot be specified together with nextSyncToken to ensure consistency of the client state. These are: - iCalUID - orderBy - privateExtendedProperty - q - sharedExtendedProperty - timeMin - timeMax - updatedMin If the syncToken expires, the server will respond with a 410 GONE response code and the client should clear its storage and perform a full synchronization without any syncToken. Learn more about incremental synchronization. Optional. The default is to return all entries.

elCalendarId :: Lens' EventsList Text

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

elTimeMin :: Lens' EventsList (Maybe UTCTime)

Lower bound (inclusive) for an event's start time to filter by. Optional. The default is not to filter by start time. Must be an RFC3339 timestamp with mandatory time zone offset, for example, 2011-06-03T10:00:00-07:00, 2011-06-03T10:00:00Z. Milliseconds may be provided but are ignored. If timeMax is set, timeMin must be smaller than timeMax.

elOrderBy :: Lens' EventsList (Maybe EventsListOrderBy)

The order of the events returned in the result. Optional. The default is an unspecified, stable order.

elSingleEvents :: Lens' EventsList (Maybe Bool)

Whether to expand recurring events into instances and only return single one-off events and instances of recurring events, but not the underlying recurring events themselves. Optional. The default is False.

elPrivateExtendedProperty :: Lens' EventsList [Text]

Extended properties constraint specified as propertyName=value. Matches only private properties. This parameter might be repeated multiple times to return events that match all given constraints.

elShowDeleted :: Lens' EventsList (Maybe Bool)

Whether to include deleted events (with status equals "cancelled") in the result. Cancelled instances of recurring events (but not the underlying recurring event) will still be included if showDeleted and singleEvents are both False. If showDeleted and singleEvents are both True, only single instances of deleted events (but not the underlying recurring events) are returned. Optional. The default is False.

elQ :: Lens' EventsList (Maybe Text)

Free text search terms to find events that match these terms in any field, except for extended properties. Optional.

elSharedExtendedProperty :: Lens' EventsList [Text]

Extended properties constraint specified as propertyName=value. Matches only shared properties. This parameter might be repeated multiple times to return events that match all given constraints.

elMaxAttendees :: Lens' EventsList (Maybe Int32)

The maximum number of attendees to include in the response. If there are more than the specified number of attendees, only the participant is returned. Optional.

elICalUId :: Lens' EventsList (Maybe Text)

Specifies event ID in the iCalendar format to be included in the response. Optional.

elUpdatedMin :: Lens' EventsList (Maybe UTCTime)

Lower bound for an event's last modification time (as a RFC3339 timestamp) to filter by. When specified, entries deleted since this time will always be included regardless of showDeleted. Optional. The default is not to filter by last modification time.

elPageToken :: Lens' EventsList (Maybe Text)

Token specifying which result page to return. Optional.

elTimeZone :: Lens' EventsList (Maybe Text)

Time zone used in the response. Optional. The default is the time zone of the calendar.

elShowHiddenInvitations :: Lens' EventsList (Maybe Bool)

Whether to include hidden invitations in the result. Optional. The default is False.

elMaxResults :: Lens' EventsList Int32

Maximum number of events returned on one result page. The number of events in the resulting page may be less than this value, or none at all, even if there are more events matching the query. Incomplete pages can be detected by a non-empty nextPageToken field in the response. By default the value is 250 events. The page size can never be larger than 2500 events. Optional.

elAlwaysIncludeEmail :: Lens' EventsList (Maybe Bool)

Whether to always include a value in the email field for the organizer, creator and attendees, even if no real email is available (i.e. a generated, non-working value will be provided). The use of this option is discouraged and should only be used by clients which cannot handle the absence of an email address value in the mentioned places. Optional. The default is False.

elTimeMax :: Lens' EventsList (Maybe UTCTime)

Upper bound (exclusive) for an event's end time to filter by. Optional. The default is not to filter by end time. Must be an RFC3339 timestamp with mandatory time zone offset, for example, 2011-06-03T10:00:00-07:00, 2011-06-03T10:00:00Z. Milliseconds may be provided but are ignored. If timeMin is set, timeMax must be greater than timeMin.