Network.Google.Resource.Calendar.Events.QuickAdd

Description

Creates an event based on a simple text string.

See: Calendar API Reference for calendar.events.quickAdd.

REST Resource

type EventsQuickAddResource = "calendar" :> ("v3" :> ("calendars" :> (Capture "calendarId" Text :> ("events" :> ("quickAdd" :> (QueryParam "text" Text :> (QueryParam "sendNotifications" Bool :> (QueryParam "sendUpdates" EventsQuickAddSendUpdates :> (QueryParam "alt" AltJSON :> Post '[JSON] Event)))))))))

A resource alias for calendar.events.quickAdd method which the EventsQuickAdd request conforms to.

Creating a Request

eventsQuickAdd

Arguments

:: Text
eqaCalendarId

-> Text
eqaText

-> EventsQuickAdd

Creates a value of EventsQuickAdd with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eqaCalendarId

eqaText

eqaSendNotifications

eqaSendUpdates

data EventsQuickAdd

Creates an event based on a simple text string.

See: eventsQuickAdd smart constructor.

Request Lenses

eqaCalendarId :: Lens' EventsQuickAdd Text

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

eqaText :: Lens' EventsQuickAdd Text

The text describing the event to be created.

eqaSendNotifications :: Lens' EventsQuickAdd (Maybe Bool)

Deprecated. Please use sendUpdates instead. Whether to send notifications about the creation of the event. Note that some emails might still be sent even if you set the value to false. The default is false.

eqaSendUpdates :: Lens' EventsQuickAdd (Maybe EventsQuickAddSendUpdates)

Guests who should receive notifications about the creation of the new event.