Network.Google.Resource.Calendar.Events.Update

Description

Updates an event.

See: Calendar API Reference for calendar.events.update.

REST Resource

type EventsUpdateResource = "calendar" :> ("v3" :> ("calendars" :> (Capture "calendarId" Text :> ("events" :> (Capture "eventId" Text :> (QueryParam "conferenceDataVersion" (Textual Int32) :> (QueryParam "maxAttendees" (Textual Int32) :> (QueryParam "sendNotifications" Bool :> (QueryParam "supportsAttachments" Bool :> (QueryParam "sendUpdates" EventsUpdateSendUpdates :> (QueryParam "alwaysIncludeEmail" Bool :> (QueryParam "alt" AltJSON :> (ReqBody '[JSON] Event :> Put '[JSON] Event)))))))))))))

A resource alias for calendar.events.update method which the EventsUpdate request conforms to.

Creating a Request

eventsUpdate

Arguments

:: Text
euCalendarId

-> Event
euPayload

-> Text
euEventId

-> EventsUpdate

Creates a value of EventsUpdate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

euConferenceDataVersion

euCalendarId

euPayload

euMaxAttendees

euSendNotifications

euSupportsAttachments

euSendUpdates

euAlwaysIncludeEmail

euEventId

data EventsUpdate

Updates an event.

See: eventsUpdate smart constructor.

Request Lenses

euConferenceDataVersion :: Lens' EventsUpdate (Maybe Int32)

Version number of conference data supported by the API client. Version 0 assumes no conference data support and ignores conference data in the event's body. Version 1 enables support for copying of ConferenceData as well as for creating new conferences using the createRequest field of conferenceData. The default is 0.

euCalendarId :: Lens' EventsUpdate Text

Calendar identifier. To retrieve calendar IDs call the calendarList.list method. If you want to access the primary calendar of the currently logged in user, use the "primary" keyword.

euPayload :: Lens' EventsUpdate Event

Multipart request metadata.

euMaxAttendees :: Lens' EventsUpdate (Maybe Int32)

The maximum number of attendees to include in the response. If there are more than the specified number of attendees, only the participant is returned. Optional.

euSendNotifications :: Lens' EventsUpdate (Maybe Bool)

Deprecated. Please use sendUpdates instead. Whether to send notifications about the event update (for example, description changes, etc.). Note that some emails might still be sent even if you set the value to false. The default is false.

euSupportsAttachments :: Lens' EventsUpdate (Maybe Bool)

Whether API client performing operation supports event attachments. Optional. The default is False.

euSendUpdates :: Lens' EventsUpdate (Maybe EventsUpdateSendUpdates)

Guests who should receive notifications about the event update (for example, title changes, etc.).

euAlwaysIncludeEmail :: Lens' EventsUpdate (Maybe Bool)

Whether to always include a value in the email field for the organizer, creator and attendees, even if no real email is available (i.e. a generated, non-working value will be provided). The use of this option is discouraged and should only be used by clients which cannot handle the absence of an email address value in the mentioned places. Optional. The default is False.

euEventId :: Lens' EventsUpdate Text

Event identifier.